refactor: split src/pecans.ts into src/http route modules (Phase 7C)
Pure move, third and final Phase 7 PR: handler bodies relocate verbatim
into src/http/{downloads,updates,api,notes}.ts against a narrow
PecansHttpContext, and the query-parsing helpers into src/http/query.ts
(re-exported from the package root, so consumer imports are unchanged).

Pecans stays the composition root: router wiring, the
beforeDownload/afterDownload events around serveAsset, and thin
delegating methods that keep the class surface (and host overrides of
public methods like queryReleases) working exactly as before. HTTP
behavior is unchanged; the contract suite passes untouched.
